change: use decimals for grid calculations (fixes lop sided grid gap)

+++ src/grid.h
@@ -1,8 +1,8 @@
 #ifndef GRID_H
 #define GRID_H
 
-unsigned int grid_x_unit_size_in_pixels;
-unsigned int grid_y_unit_size_in_pixels;
+float grid_x_unit_size_in_pixels;
+float grid_y_unit_size_in_pixels;
 
 extern unsigned int grid_window_default_x;
 extern unsigned int grid_window_default_y;
@@ -11,13 +11,13 @@ extern unsigned int grid_window_default_width;
 
 void apply_configuration_to_grid(void);
 
-unsigned int grid_get_span(unsigned int, unsigned int);
-unsigned int grid_get_offset(unsigned int, unsigned int);
+float grid_get_span(float, unsigned int);
+float grid_get_offset(float, unsigned int);
 
-unsigned int grid_get_span_x(unsigned int);
-unsigned int grid_get_span_y(unsigned int);
+float grid_get_span_x(unsigned int);
+float grid_get_span_y(unsigned int);
 
-unsigned int grid_get_x_offset(unsigned int);
-unsigned int grid_get_y_offset(unsigned int);
+float grid_get_x_offset(unsigned int);
+float grid_get_y_offset(unsigned int);
 
 #endif /* GRID_H */
